+++ Game1/Shapes/Ellipse.cs
@@ -15,8 +15,8 @@
         public override bool Contains(MyVector2 position)
         {
             MyVector2 relPos = position - Center;
-            double propX = 2 * relPos.X / (double)Width,
-                propY = 2 * relPos.Y / (double)Height;
+            double propX = 2 * relPos.X / Width,
+                propY = 2 * relPos.Y / Height;
             return propX * propX + propY * propY < 1;
         }
